pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robotframework-testselection"
dynamic = ["version"]
description = "Vector-based diverse test case selection for Robot Framework"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Many Kasiriha" },
]
keywords = [
    "robotframework",
    "testing",
    "test-selection",
    "diversity",
    "machine-learning",
    "nlp",
    "embeddings",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Robot Framework",
    "Framework :: Robot Framework :: Library",
    "Framework :: Robot Framework :: Tool",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "robotframework>=7.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.urls]
Homepage = "https://github.com/manykarim/robotframework-testselection"
Documentation = "https://github.com/manykarim/robotframework-testselection#readme"
Repository = "https://github.com/manykarim/robotframework-testselection"
Issues = "https://github.com/manykarim/robotframework-testselection/issues"
Changelog = "https://github.com/manykarim/robotframework-testselection/blob/main/CHANGELOG.md"

[project.optional-dependencies]
vectorize = [
    "sentence-transformers>=2.2",
]
selection-extras = [
    "scikit-learn-extra>=0.3",
    "dppy>=0.3",
    "apricot-select>=0.6",
]
chromadb = [
    "chromadb>=0.4",
]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "pytest-benchmark",
    "ruff",
    "mypy",
]
all = [
    "robotframework-testselection[vectorize,selection-extras,chromadb,dev]",
]

[project.scripts]
testcase-select = "TestSelection.cli:main"

[tool.hatch.version]
path = "src/TestSelection/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/TestSelection"]

[tool.hatch.build.targets.sdist]
only-include = [
    "src/TestSelection",
    "LICENSE",
    "README.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "slow: marks tests that load ML models or large datasets",
    "integration: marks integration tests requiring full pipeline",
    "benchmark: marks benchmark tests",
]

[tool.ruff]
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = false
packages = ["TestSelection"]
mypy_path = "src"
